add string sanitizer for HTML
